Added BLITFIX + minor bugfixes in all programs
BLITFIX tries to allow the blitter as much as possible, while preventing its use outside of the 24bit address range.
Mainly fixes so TOS 2.06 can be used together with TT-RAM, and also NVDI.
Small benefits also to EmuTOS, allowing programs in ST-RAM to get correct results when asking if blitter is present.